Методы по подготовке изображения вынесены в отдельный временный класс
